requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
